1) The spellcheck thing is coming on because in order to use the execCommand, you have to turn on "designMode", which turns the entire page into editable content (and thus turns on spellcheck). Then, we move our selection to our word, and this triggers spellcheck on the word. So, ideally you would just turn on designMode after you've moved your selection so it won't trigger spellcheck, highlight, and then turn off designmode--and this does work to eliminate this issue--but it's like 2X slower, and this lag is noticeable.
So none of the easy options are good. That is, easy option 1 is to have a bunch of red squiggly lines everywhere and option 2 is make it go 2X slower, none of which are good. So i attempted to re-code execCommand myself so as to not require designmode. This roughly equated to reverting the highlighting technique back to the old way you had things, but with some new tricks. I ended up getting really, really close, but running into one or two unexplainable errors dealing with ranges and I just can't figure out why the code fails here but doesn't there, etc. And I never really got to the hard part of dealing with a selection that spans multiple nodes (such as selecting both a bolded word and an unbolded one next to it). So, after too many hours of slogging through the stuff, I'm giving up, and personally taking easy option 2.

Here's what i've ended up with as far as using execCommand: http://jsfiddle.net/uS6b6/9/
The pros are that it searches textareas and totally works with selecting spliced elements. The cons are that it is significantly slower... it can get significantly faster by removing the toggle on designmode, but then that introduces red squiggly lines everywhere.

Unless something new comes into the picture from your end, I think this is as far as I can take it for now. I really wish chrome had more resources on seeing how execCommand backcolor works, how to deal with its spellcheck, or even some kind of way to trigger its ctrl+f  functionality.